requirements.txt
Flask==2.3.2
google-cloud-texttospeech==2.14.2